moved tabs to lab components
