+++ src/lab/Tabs/styles.ts
